Showcasing innovative UI/UX design for healthcare communication.
Curecoom is a UI/UX design project that reflects a deep understanding of the needs of individuals seeking health-related community support. This repository presents a series of Figma designs that demonstrate a user-centric approach to creating intuitive and engaging interfaces for a healthcare communication app. The designs emphasize ease of use, accessibility, and the facilitation of meaningful connections among users with similar health conditions